Move hitzone and reward condition entities to a common package
- Move `HitzoneEntity`, `HitzoneTextEntity`, `RewardConditionEntity`, and `RewardConditionTextEntity` from `monster` to `common` package.
- Update imports and references.
