+++ core/data/src/main/java/com/gaugustini/mhfudatabase/data/database/entity/monster/MonsterHitzoneEntity.kt
@@ -4,6 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
+import com.gaugustini.mhfudatabase.data.database.entity.common.HitzoneEntity
 
 @Entity(
     tableName = "monster_hitzone",
