+++ core/data/src/main/java/com/gaugustini/mhfudatabase/data/database/entity/quest/QuestRewardEntity.kt
@@ -4,8 +4,8 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
+import com.gaugustini.mhfudatabase.data.database.entity.common.RewardConditionEntity
 import com.gaugustini.mhfudatabase.data.database.entity.item.ItemEntity
-import com.gaugustini.mhfudatabase.data.database.entity.monster.RewardConditionEntity
 
 @Entity(
     tableName = "quest_reward",
