+++ core/data/src/main/java/com/gaugustini/mhfudatabase/data/database/relation/MonsterRewardWithMonster.kt
@@ -1,10 +1,10 @@
 package com.gaugustini.mhfudatabase.data.database.relation
 
 import androidx.room.Embedded
+import com.gaugustini.mhfudatabase.data.database.entity.common.RewardConditionTextEntity
 import com.gaugustini.mhfudatabase.data.database.entity.monster.MonsterEntity
 import com.gaugustini.mhfudatabase.data.database.entity.monster.MonsterRewardEntity
 import com.gaugustini.mhfudatabase.data.database.entity.monster.MonsterTextEntity
-import com.gaugustini.mhfudatabase.data.database.entity.monster.RewardConditionTextEntity
 
 /**
  * Represents a monster reward entity with its associated condition, monster, and text entities.
